Extra loginTest was deleted
