+++ morozovaProj/src/test/java/loginTests/LoginTest.java
@@ -1,6 +1,7 @@
 package loginTests;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 import org.openqa.selenium.By;
@@ -15,7 +16,7 @@ public class LoginTest {
    WebDriver webDriver;
 
    @Test
-   public void validLoginIn(){
+   public void validLoginIn() {
       WebDriverManager.chromedriver().setup();//maven zarune chromedriver
       webDriver = new ChromeDriver();
       webDriver.manage().window().maximize();
@@ -34,11 +35,53 @@ public class LoginTest {
       System.out.println("Password was inputted");
       webDriver.findElement(By.xpath(".//button[@class='btn btn-primary btn-sm']")).click();
       System.out.println("Button was clicked");
-      WebElement buttonSignOut = webDriver.findElement(By.xpath(".//button[@class='btn btn-primary btn-sm']"));
-      Assert.assertTrue("Button is not displayed", buttonSignOut.isDisplayed());
+//        WebElement buttonSignOut = webDriver.findElement(By.xpath(".//button [text()= 'Sign Out']"));
+      Assert.assertTrue("Button is not displayed", isButtonSignOutVisible());
 
-      webDriver.quit();
-      System.out.println("Browser was closed");
+//        webDriver.quit();
+//        System.out.println("Browser was closed");
 
    }
+   @After //this method will be executed after each test
+   public void tearDown() {
+      webDriver.quit();
+      System.out.println("Browser was closed");
+   }
+
+
+   private boolean isButtonSignOutVisible() {
+      try {
+         return webDriver.findElement(By.xpath(".//button [text()= 'Sign Out']")).isDisplayed();
+      } catch (Exception e) {
+         return false;
+      }
+   }
+//    @Test
+//public void invalidLoginIn() {
+//        WebDriverManager.chromedriver().setup();//maven zarune chromedriver
+//        webDriver = new ChromeDriver();
+//        webDriver.manage().window().maximize();
+//        webDriver.manage().timeouts().implicitlyWait(55, TimeUnit.SECONDS);
+//        System.out.println("Browser was opened");
+//        webDriver.get("https://qa-complexapp.onrender.com");
+//        System.out.println("Site was opened");
+//        WebElement inputUserName =
+//                webDriver.findElement(By.xpath("//input[@placeholder='Username']"));
+//        inputUserName.clear();
+//        inputUserName.sendKeys("qaauto");
+//        System.out.println("Username was inputted");
+//        WebElement inputPassword = webDriver.findElement(By.xpath("//input[@placeholder='Password']"));
+//        inputPassword.clear();
+//        inputPassword.sendKeys("123456qwerty");
+//        System.out.println("Password was inputted");
+//        webDriver.findElement(By.xpath(".//button[text()='Sign In']")).click();
+//        System.out.println("Button was clicked");
+//        WebElement buttonSignOut = webDriver.findElement(By.xpath(".//button [text()= 'Sign Out']"));
+//        Assert.assertTrue("Button is not displayed", isButtonSignOutVisible());
+//
+//        webDriver.quit();
+//        System.out.println("Browser was closed");
+//
+//    }
+
 }
